[refactor]: split bottom sheet widget tree
