+++ lib/main.dart
@@ -73,7 +73,7 @@ class DalalApp extends StatelessWidget {
               if (state.user.isPhoneVerified) {
                 //showSnackBar(context, 'Welcome ${state.user.name}');
                 _navigator.pushNamedAndRemoveUntil(
-                  '/exchange',
+                  '/home',
                   (route) => false,
                   arguments: state.user,
                 );
